requirements.txt
openai>=2.0.0
datasets>=4.0.0
pandas>=2.0.0
numpy>=1.23.0
scipy>=1.10.0
PyYAML>=6.0
tqdm>=4.60.0
nest_asyncio>=1.5.0
prettytable>=3.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
ipykernel>=6.0.0
